Added db.Model inheritance, tested first __repr__, moved connect_to_db to server.py
